test(vmware-explorer): separate the tests which need sockets, cover the review findings
