+++ @xen-orchestra/vmware-explorer/_thumbprint.integ.mjs
